requirements.txt
python-telegram-bot==20.3
flask==2.2.3
flask-cors==3.0.10
cryptography==39.0.1